Make buildabe with gcc

+++ src/runtime/ag-queue.c
@@ -6,8 +6,8 @@ void ag_init_queue(ag_queue* q) {
 	q->end = q->start + AG_THREAD_QUEUE_SIZE;
 }
 
-void ag_resize_queue(ag_queue* q, size_t space_needed) {
-	size_t free_space = q->read_pos > q->write_pos
+void ag_resize_queue(ag_queue* q, uint64_t space_needed) {
+	uint64_t free_space = q->read_pos > q->write_pos
 		? q->write_pos - q->read_pos
 		: (q->end - q->start) - (q->write_pos - q->read_pos);
 	if (free_space < space_needed) {
@@ -16,15 +16,15 @@ void ag_resize_queue(ag_queue* q, size_t space_needed) {
 		if (!new_buf)
 			exit(-42);
 		if (q->read_pos > q->write_pos) {
-			size_t r_size = q->end - q->read_pos;
-			size_t w_size = q->write_pos - q->start;
+			uint64_t r_size = q->end - q->read_pos;
+			uint64_t w_size = q->write_pos - q->start;
 			memcpy(new_buf + new_size - r_size, q->read_pos, sizeof(uint64_t) * (r_size));
 			memcpy(new_buf, q->start, sizeof(uint64_t) * (w_size));
 			AG_FREE(q->start);
 			q->read_pos = new_buf + new_size - r_size;
 			q->write_pos = new_buf + w_size;
 		} else {
-			size_t size = q->write_pos - q->read_pos;
+			uint64_t size = q->write_pos - q->read_pos;
 			memcpy(new_buf, q->read_pos, sizeof(uint64_t) * size);
 			AG_FREE(q->start);
 			q->read_pos = new_buf;
